Create README.md
This study presents the impact of individuals on sustainable economic growth through digital technology. This presentation describes a how mobile e-learning (Digital Solutions) brings many unprivileged women and youth closer to accessing the internet and an opportunity to broaden their knowledge and skills in their endeavors. It also describes the already existing interventions and how this solution brings guidance to individuals who do not know the path to digitalization and its importance to them and the country as a whole.